+++ geocat/comp/spherical.py
@@ -1,9 +1,23 @@
 from typing import Union
 
 import numpy as np
-import scipy.special as sspecial
-import scipy.spatial as sspatial
+from packaging.version import Version
+from scipy.spatial import SphericalVoronoi
+from scipy import __version__ as scipy_version
 import xarray as xr
+
+# import scipy shp_harm[_y] function depending on scipy version
+old_scipy = False
+scipy_version = Version(scipy_version)
+if scipy_version < Version('1.15.0'):
+    from scipy.special import sph_harm
+
+    def sph_harm_y(n, m, theta, phi):
+        return sph_harm(m, n, phi, theta)
+
+    old_scipy = True
+else:
+    from scipy.special import sph_harm_y
 
 SupportedTypes = Union[np.ndarray, xr.DataArray]
 default_max_harm = 23  # 300 harmonics from 0,0 to 23,23
@@ -15,7 +29,6 @@ def decomposition(
     theta: SupportedTypes,
     phi: SupportedTypes,
     max_harm: int = default_max_harm,
-    chunk_size: int = 'auto',
 ) -> SupportedTypes:
     """Calculate the spherical harmonics of a dataset.
 
@@ -41,26 +54,22 @@ def decomposition(
         The total of harmonics calculated is ``(max_harm+1)*(max_harm+2)/2``
         Defaults to 23, for 300 total harmonics.
 
-    chunk_size: int, optional
-        The size of each edge of the dask chunks if using xarray.DataArray inputs.
-        Some arrays will be 2d, and others 1d, and the final calculation operates on a 3d array.
-        thus the chunks used in the largest calculation scale at ``chunk_size^3``
-        A chunk size of 256 is recommended. Defaults to 'auto'
-
     Returns
     -------
     decomposition : ndarray, :class:`xarray.DataArray`
         The spherical harmonic decomposition of the input data
     """
 
+    in_type = type(data)
+
     # scale_val is the inverse of the total sphere area times the magnitude of
     # the first harmonic. This is used to scale the output so that the output
     # is unaffected by the surface area of the original sphere.
-    scale_val = 1 / (np.sum(scale, axis=(0, 1)) *
-                     sspecial.sph_harm(0, 0, 0, 0)**2)
+    scale_val = np.array(
+        (1 / (np.sum(scale, axis=(0, 1)) * sph_harm_y(0, 0, 0, 0)**2)))
 
-    mlist = []  # ordered list of the m harmonics sspecial.sphere(m,n,theta,phi)
-    nlist = []  # ordered list of the n harmonics sspecial.sphere(m,n,theta,phi)
+    mlist = []  # ordered list of the m harmonics
+    nlist = []  # ordered list of the n harmonics
     # the real value of the output varies on a factor of two when m is 0,
     # due to m=0 being a symmetric case with no imaginary component,
     # this accounts for that difference in output sum.
@@ -73,38 +82,44 @@ def decomposition(
                 scale_mul.append(1)
             else:
                 scale_mul.append(2)
-    m = np.array(mlist)
-    n = np.array(nlist)
+    m = np.expand_dims(mlist, axis=(1, 2))
+    n = np.expand_dims(nlist, axis=(1, 2))
     scale_mul = np.array(scale_mul)
+    scale_res = scale_mul * scale_val
+    scale_dat = np.multiply(data, scale)
 
-    # if numpy, change dimensions to allow for broadcast in ss.sph_harm
-    if type(data) is np.ndarray:
-        m = np.expand_dims(m, axis=(0, 1))
-        n = np.expand_dims(n, axis=(0, 1))
-        theta = np.expand_dims(theta, axis=(2))
-        phi = np.expand_dims(phi, axis=(2))
-        scale_res = scale_mul * scale_val
-        scale_dat = np.expand_dims(np.multiply(data, scale), axis=(2))
+    # set dims for broadcasting
+    if in_type is xr.DataArray:
+        if not old_scipy:
+            scale_res = xr.DataArray(scale_res, dims=['harmonic'])
+            scale_dat = xr.DataArray(scale_dat, dims=data.dims)
+            theta = xr.DataArray(theta, dims=data.dims)
+            phi = xr.DataArray(phi, dims=data.dims)
+        else:
+            theta = np.expand_dims(theta, axis=0)
+            phi = np.expand_dims(phi, axis=0)
 
-    # if xarray, set dims and chunks for broadcast in ss.sphere_harm
-    if type(data) is xr.DataArray:
-        m = xr.DataArray(m, dims=['har']).chunk((chunk_size))
-        n = xr.DataArray(n, dims=['har']).chunk((chunk_size))
-        scale_res = xr.DataArray(
-            scale_mul,
-            dims=['har'],
-        ).chunk((chunk_size)) * scale_val
-        scale_dat = xr.DataArray(
-            np.multiply(data, scale),
-            dims=data.dims,
-        ).chunk((chunk_size))
-        theta = xr.DataArray(theta, dims=data.dims).chunk((chunk_size))
-        phi = xr.DataArray(phi, dims=data.dims).chunk((chunk_size))
+        # np.multiply produces (lat, lon, harmonic) dimensions for xarray input
+        sum_ax = (0, 1)
+    else:
+        theta = np.expand_dims(theta, axis=0)
+        phi = np.expand_dims(phi, axis=0)
 
-    results = np.sum(
-        np.multiply(scale_dat, sspecial.sph_harm(m, n, theta, phi)),
-        axis=(0, 1),
-    ) * scale_res
+        # np.multiply produces (harmonic, lat, lon) dimensions for numpy input
+        sum_ax = (1, 2)
+
+    # scipy 1.15 flips the definitions of theta and phi
+    theta, phi = phi, theta
+
+    harmonics = sph_harm_y(n, m, theta, phi)
+
+    # if xarray, make harmonics into xarray and align dims
+    if in_type is xr.DataArray:
+        harmonics = xr.DataArray(harmonics,
+                                 dims=['harmonic', data.dims[0], data.dims[1]])
+
+    results = np.sum(np.multiply(scale_dat, harmonics), axis=sum_ax) * scale_res
+
     return results
 
 
@@ -113,7 +128,6 @@ def recomposition(
     theta: SupportedTypes,
     phi: SupportedTypes,
     max_harm: int = default_max_harm,
-    chunk_size: int = 'auto',
 ) -> SupportedTypes:
     """Calculate a dataset from spherical harmonics.
 
@@ -136,50 +150,50 @@ def recomposition(
         The number of total harmonics must equal the number of harmoncs in the input data.
         Defaults to 23, for 300 total harmonics.
 
-    chunk_size: int, optional
-        The size of each edge of the dask chunks if using xarray.DataArray inputs.
-        Some arrays will be 2d, and others 1d, and the final calculation operates on a 3d array.
-        thus the chunks used in the largest calculation scale at ``chunk_size^3``
-        A chunk size of 256 is recommended. Defaults to 'auto'
-
     Returns
     -------
     recomposition : ndarray, :class:`xarray.DataArray`
         The spherical harmonic recomposition of the input data
     """
 
-    mlist = []  # ordered list of the m harmonics sspecial.sphere(m,n,theta,phi)
-    nlist = []  # ordered list of the n harmonics sspecial.sphere(m,n,theta,phi)
+    in_type = type(data)
+
+    # if xarray, standardize data input dimension names
+    if in_type is xr.DataArray:
+        data = data.rename({data.dims[0]: 'harmonic'})
+        theta_dims = theta.dims
+
+    mlist = []  # ordered list of the m harmonics
+    nlist = []  # ordered list of the n harmonics
     for nvalue in range(max_harm + 1):
         for mvalue in range(nvalue + 1):
             mlist.append(mvalue)
             nlist.append(nvalue)
-    m = np.array(mlist)
-    n = np.array(nlist)
 
-    # if numpy, change dimensions to allow for broadcast in ss.sph_harm
-    if type(data) is np.ndarray:
-        m = np.expand_dims(m, axis=(1, 2))
-        n = np.expand_dims(n, axis=(1, 2))
+    m = np.expand_dims(mlist, axis=(1, 2))
+    n = np.expand_dims(nlist, axis=(1, 2))
+
+    # set dims for broadcasting
+    if in_type is xr.DataArray and not old_scipy:
+        theta = theta.expand_dims(dim='harmonic', axis=0)
+        phi = phi.expand_dims(dim='harmonic', axis=0)
+    else:
         data = np.expand_dims(data, axis=(1, 2))
-        theta = np.expand_dims(theta, axis=(0))
-        phi = np.expand_dims(phi, axis=(0))
+        theta = np.expand_dims(theta, axis=0)
+        phi = np.expand_dims(phi, axis=0)
 
-    # if xarray, set dims and chunks for broadcast in ss.sphere_harm
-    if type(data) is xr.DataArray:
-        m = xr.DataArray(m, dims=['har']).chunk((chunk_size))
-        n = xr.DataArray(n, dims=['har']).chunk((chunk_size))
-        data = xr.DataArray(data, dims=['har']).chunk((chunk_size))
-        theta = xr.DataArray(theta, dims=theta.dims).chunk((chunk_size))
-        phi = xr.DataArray(phi, dims=phi.dims).chunk((chunk_size))
+    # scipy 1.15 flips the definitions of theta and phi
+    theta, phi = phi, theta
 
-    results = np.sum(
-        np.multiply(sspecial.sph_harm(m, n, theta, phi).real, data.real),
-        axis=(0),
-    ) + np.sum(
-        np.multiply(sspecial.sph_harm(m, n, theta, phi).imag, data.imag),
-        axis=(0),
-    )
+    harmonics = sph_harm_y(n, m, theta, phi)
+
+    # if xarray, make harmonics into xarray and align dims
+    if in_type is xr.DataArray:
+        harmonics = xr.DataArray(
+            harmonics, dims=['harmonic', theta_dims[0], theta_dims[1]])
+
+    results = (np.sum(np.multiply(harmonics.real, data.real), axis=0) +
+               np.sum(np.multiply(harmonics.imag, data.imag), axis=0))
 
     return results.real
 
@@ -225,7 +239,7 @@ def scale_voronoi(
     data_locs_3d[:, 2] = np.cos(phi_1d)
 
     scale = np.array(
-        sspatial.SphericalVoronoi(
+        SphericalVoronoi(
             data_locs_3d,
             radius=1.0,
             center=np.array([0, 0, 0]),